New default values showing non-profitability
The proposed example is somewhat extreme: An attacker with 20% of the
mining power attacks and steals 128 BTC. At a discount rate of 0.9995,
this attack is not profitable. However, at a discount rate of 1, the
attack becomes profitable in the future (once the exchange rate, and
hence the value of mining ASICs, recovers).